generalization f1 second attempt, before log-spaced lambdas

+++ codes/deploy_compare.py
@@ -25,7 +25,7 @@ from trajectories import heldout_set, custom_set, showcase_set
 # SHOWCASE: net-vs-optimizer overlay on ONE demo trajectory (decoupled from the training set) at
 # its OWN horizon — shows the net handles any length. None | "short" | "long" (trajectories.
 # showcase_set); SHOWCASE_IDX picks which entry (0=line, 1..=quintics). Overrides the toggles below.
-SHOWCASE = "long"
+SHOWCASE = "short"
 SHOWCASE_IDX = 2      # 0 -> the straight-line demo; 1.. -> the quintic demos
 SHOWCASE_M = 3        # quintics available per preset (so SHOWCASE_IDX can reach 1..M)
